requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
cryptography==41.0.4
Werkzeug==2.3.7
python-dotenv==1.0.0
reportlab==4.0.4
Pillow==10.0.0